+++ src/pykeen/sampling/__init__.py
@@ -116,7 +116,7 @@ from class_resolver import Resolver, get_subclasses
 from .basic_negative_sampler import BasicNegativeSampler
 from .bernoulli_negative_sampler import BernoulliNegativeSampler
 from .negative_sampler import NegativeSampler
-from .pseudo_type_negative_sampling import PseudoTypedNegativeSampler
+from .pseudo_type import PseudoTypedNegativeSampler
 
 __all__ = [
     'NegativeSampler',
